rag-iep-system/services/upsert-function/requirements.txt
google-cloud-aiplatform==1.38.0
functions-framework==3.5.0
